Parallelize drive API
This adds concurrent functionality to the Google Drive API library.
Previously, we would fetch the files and folders starting from the
root folder one at a time. With this update, the app will
concurrently fetch all files and folders within a directory all at
once and the application will keep track of the results in thread
safe arrays.