Increase drive api effeciency
This commit updates the way we pull file metadata down from Google
drive. Instead of starting at the root folder and recursively
traversing the file tree, we instead pull all the files down
locally, including the parent_id, and then build the full file path
on the local server using the parent_ids. This is way more
efficient because we are able to fetch all of the files in less
than 10 API calls versus one API call per directory, which is over
a hundred. This is slightly more work for the local server, but
that's fine because we aren't strapped for resources and the amount
of compute needed is barely any higher than what we were doing
before.