Memoize variables in file tree differ
This commit adds memoization to the missing folder methods in the
file tree differ class so that we don't need to recalculate
everytime the method is called. With that change, we can simplify
the diff method so that we don't have to keep track of the missing
folders in a local variable.

We also rename the methods to remove the work get_ since that
violates rubocop's best practices.