Add file metadata classes and refactor differ
This commit overhauls the application quite a bit. We introduce two
new classes that encapsulate the information and logic for the file
metadata. We then refactor the file_tree_differ class to work with
these new classes.

This also refactors the spec tests for the differ class as well as
adds spec tests for the new file_metadata class.