Extract drive service initializer, add local sync
This commit extracts out the logic to initialize the drive service
into a separate class. This is best practice to allow for
dependency injection, which makes the code easily testable.

We also add the functionality to download file contents in the
drive API. With this functionality, we can now sync missing files
from Google Drive. This option is enabled using the -l flag.

One final change we made is to use optional arguments for running
the program. Without the arguments, this app will remain unchanged
in that it will still diff and report missing files between my
backup drive and Google Drive documents folder.