Add summarize_printout option to file tree differ
This adds a new param to the FileTreeDiffer class to specify
whether or not the diff print output should summarize directories
or print every individual file. For example, if the follow is
missing:
- documents/folder
- documents/folder/file1.txt

So the folder isn't even present on one of the volumes, then
summarize_printout enabled will only print that the entire
directory is missing and not print the files.

We also add complete YARD documentation to the FileTreeDiffer
class.