Add summarized-printout option, refactor unsynced-list
This adds a new option to control whether or not to summarize the
diffs printout. This specifies whether the diff print output should
summarize directories or print every individual file.
For example, if the following is missing:
  - documents/folder
  - documents/folder/file1.txt

In this example, the folder isn't present on one of the volumes
and when no-summarize-printout is set, the application
will print that the directory is missing as well as print the
file. If this flag is not set, summarization is enabled which would
cause the printout to only display the directory which is easier to
understand that the entire directory is missing.

The default value for the unsynced-list option is removed. If the
-u flag is unset, the application will not load an unsynced list
which indicates all files need to be synced between volumes.