Add tests for main and extract options into method
This adds full spec test coverage for the main class. In an effort
to get full score from Simplecov for test coverage, I extracted the
options parsing into a method so that it could be tested. I learned
how to skip Simplecov with annotations after I wrote the tests for
the options, so im just going to leave it. However, we do skip
coverage checks on the file name check.

The filename check is necessary because when the main.rb file was
required as an import in the spec file, it would execute the
script. This check ensures that the code isn't executed if the file
is required. It now needs to be explicitly called.