Add Dockerfile build configurations
This commit adds a Dockerfile to build a deployable image. We also
add a .dockerignore file and strictly limit what gets copied into
the image based on necessity.

We also add a new script to build and push to our container
registry. It requires a .env file to set the registry username.
For that reason, we also updates the .gitignore to ignore any
.env files.