Add pull_and_execute script
Adds a pull_and_execute script that will be utilized on the
production server to pull the latest image, set the correct env
variables from a .env file, and then execute the image in a
container.

I opted to write the logic for the timeout so that the host machine
doesn't need to have coreutils installed to utilize the bash timeout
function.